+++ src/App.tsx
@@ -2,6 +2,7 @@ import { HashRouter, Routes, Route, useNavigate } from 'react-router-dom'
 import { useEffect } from 'react'
 import { Toaster } from 'sonner'
 import { ThemeProvider, useTheme } from './contexts/ThemeContext'
+import { NavigationHistoryProvider } from './contexts/NavigationHistoryContext'
 import { AppLayout } from './components/AppLayout'
 import { HomePage } from './pages/HomePage'
 import { ProfilePage } from './pages/ProfilePage'
@@ -42,20 +43,22 @@ export default function App() {
   return (
     <ThemeProvider defaultTheme="dark">
       <HashRouter>
-        <ElectronNavigator />
-        <ThemedToaster />
-        <Routes>
-          <Route element={<AppLayout />}>
-            <Route path="/" element={<HomePage />} />
-            <Route path="/profile" element={<ProfilePage />} />
-            <Route path="/settings" element={<SettingsPage />} />
-            <Route path="/trash" element={<TrashPage />} />
-            <Route path="/activities" element={<ActivitiesPage />} />
-            <Route path="/activities/new" element={<ActivityFormPage />} />
-            <Route path="/activities/:id" element={<ActivityDetailPage />} />
-            <Route path="/activities/:id/edit" element={<ActivityFormPage />} />
-          </Route>
-        </Routes>
+        <NavigationHistoryProvider>
+          <ElectronNavigator />
+          <ThemedToaster />
+          <Routes>
+            <Route element={<AppLayout />}>
+              <Route path="/" element={<HomePage />} />
+              <Route path="/profile" element={<ProfilePage />} />
+              <Route path="/settings" element={<SettingsPage />} />
+              <Route path="/trash" element={<TrashPage />} />
+              <Route path="/activities" element={<ActivitiesPage />} />
+              <Route path="/activities/new" element={<ActivityFormPage />} />
+              <Route path="/activities/:id" element={<ActivityDetailPage />} />
+              <Route path="/activities/:id/edit" element={<ActivityFormPage />} />
+            </Route>
+          </Routes>
+        </NavigationHistoryProvider>
       </HashRouter>
     </ThemeProvider>
   )